Updated:
Individual Workflow
- To check probation based on the start_Date of the user and the current date of application. Will calculate based on the difference in number of days set in workflow criteria

- Notification table to include a date of approval as well as a remarks column

TRF/TNF UI Side
- Auto populated user information based on login session
- Redirection of user to login if no session is detected in applyCourse.aspx

DB Schema
- Updated the DB Schema to reflect the changes in the TRF Form fields, User table to include a start_date, notification table to include a dateApproved and remarks

+++ app/LearnHub/LearnHub/AppCode/dao/WorkflowDAO.cs
@@ -31,7 +31,7 @@ namespace LearnHub.AppCode.dao
                 {
                     toReturn = new Workflow();
                     toReturn.setWorkflowID((int)dr["wfid"]);
-                    toReturn.setProbationPeriod((double)(dr["probation_period"]));
+                    toReturn.setProbationPeriod((int)(dr["probation_period"]));
                     toReturn.setBondCriteria((double)(dr["bond_criteria"]));
                     string userCreatedID = (string)dr["userID"];
                     toReturn.setUserCreated(userDAO.getUserByID(userCreatedID));
@@ -68,7 +68,7 @@ namespace LearnHub.AppCode.dao
                 {
                     toReturn = new Workflow();
                     toReturn.setWorkflowID((int)dr["wfid"]);
-                    toReturn.setProbationPeriod((double)(dr["probation_period"]));
+                    toReturn.setProbationPeriod((int)(dr["probation_period"]));
                     toReturn.setBondCriteria((double)(dr["bond_criteria"]));
                     string userCreatedID = (string)dr["userID"];
                     toReturn.setUserCreated(userDAO.getUserByID(userCreatedID));
